# Major speedup of floating-point handling in JSON/Text