--- a/Sources/SwiftProtobuf/MathUtils.swift
+++ b/Sources/SwiftProtobuf/MathUtils.swift
@@ -12,12 +12,7 @@
 ///
 // -----------------------------------------------------------------------------
 
-#if os(Linux)
-// Linux doesn't seem to define these by default.
-// https://bugs.swift.org/browse/SR-4198
-internal let FLT_DIG=6
-internal let DBL_DIG=15
-#endif
+import Foundation
 
 /// Remainder in standard modular arithmetic (modulo). This coincides with (%)
 /// when a > 0.