Previously, Float and Double handling for JSON/Text were handled with
the `Double(String)`, `Float(String)`, `String(Float)`, and
`String(Double)` initializers.  This was very slow, with a lot
of time spent in constructing String temporaries and translating those
temporaries to and from UTF8.

To decode a float, for instance, we would convert the UTF8 into
a String, then the Swift Standard Library would convert the string
back into UTF8 in order to use the C library `strtof` function
internally.

This code cuts out the middleman: We pass the UTF8 bytes directly to
the C library functions `strtod` and `vsnprintf` ourselves, avoiding
the intermediate String objects entirely.

For decode, this is a 10x improvement, putting float/double
parsing much more nearly on par with integer handling.

For encode, this is a big (4x) improvement, but we still have
a long way to go.  Invoking `vsnprintf` from Swift requires using
Swift's expensive `CVarArg` facility; profiling shows about 50% of
our total time being spent just to marshal varargs.

Actual results: After modifying the `perf_runner` test to write out
numbers with fractional parts, I now see the following results for
encoding/decoding 100 float fields to/from Text format (JSON results
are similar):

* Before: 600 µs Text format encoding, 260 µs decoding
* After: 140 µs Text format encoding, 22 µs decoding
* C++: 45 µs Text format encoding, 36 µs decoding

For comparison, encoding 100 int64 fields using the SwiftProtobuf text
encoder takes about 12µs on the same hardware.