@allevato 2017-03-18
Hoping this doesn't sound too critical—

Making things faster is a good goal, but it feels like we're going even further down the path of "assembly language with Swift syntax" that I mentioned before with these changes. Are we sacrificing readability and maintainability of the runtime library by doing this? Are we going to be able to look at this code in 6 months and understand it easily?

When I made the original JSON performance boost, it was still very much idiomatic Swift—it used the high-level string view APIs, indexes, and so forth—but as time goes on, more and more of the runtime library looks less and less like Swift. The readability and design of the code has taken a distant back seat to squeezing out extra microseconds, and I wonder if we need to rebalance these priorities.

For binary encoding, I'm willing to concede that dropping into unsafe-land to squeeze out performance gains is probably something we need to do, since the whole point of the encoding is to be small and fast. But for text and JSON, if we've reached the point where we're mostly passing around raw UTF-8 pointers and calling into C functions, what's the point of writing it in Swift at all? We could get even better performance by implementing the whole thing in C and wrapping it in a Swift interface. _(Obviously I don't want to do this.)_

Another fear I have is that we're setting the efficiency bar impossibly high, and when the time comes that we want to revisit parts of the runtime design and improve them, we'll hesitate because they might result in a slight runtime increase in the raw numbers, even if those numbers would still be "great" by an objective standard.

We're already matching or beating C++ in text/JSON performance metrics, so is it worth going farther beyond that when we end up paying for it in code readability?