@tbkka 2017-03-20
> Hoping this doesn't sound too critical...

Although motivated by performance, this particular PR is also a step towards developing UTF-8-oriented support code that I think will make the JSON and Text logic easier to understand.  Specifically, I've created a `DoubleFormatter` that hides details of converting between UTF-8 and Float/Double.  Isolating those details made it feasible to explore using C library functions that happen to be much better suited to UTF-8-oriented operations.  My description above explains this in holistic terms:  How the compiled form now works after everything gets optimized down.  But I think the source is actually clearer than before.

Swift's standard String support is heavily biased towards UTF-16 and has a lot of compromises in its UTF-8 handling right now.  JSON and Text formatting and parsing are inherently UTF-8 based, and it's not yet entirely clear to me how best to express UTF-8 operations in Swift.  Today, we're using a mix of techniques: JSON/Text encoding relies heavily on `Array<UInt8>`, while decoding is using simple pointers a lot.  Two easy improvements I hope to work on soon:  migrating a lot of the simple pointers to buffer pointers, and developing tailored types to wrap key concepts (`NameMap.Name` and `DoubleFormatter` are two steps in that direction).  I expect that future Swift evolution will provide even better tools.

> We could get even better performance by implementing the whole thing in C ... (Obviously I don't want to do this.)

Nor do I.  But I'm not averse to selectively using small bits of C where C APIs happen to be a better fit for the task at hand.  Parsing floats and doubles from UTF-8 is one of those cases; `strtod` is almost exactly the right API for that.  Formatting floats and doubles into UTF-8 may be another, though `vsnprintf` is not the most elegant API around, and the `CVarArgs` tax is higher than I expected.  At least the current PR manages to isolate that operation so we can continue looking for better ways to express it.

> Another fear I have is that we're setting the efficiency bar impossibly high...

My first focus for Swift Protobuf was correctness.  As I learned what that involved, I wrote a lot of ugly code filled with special cases that eventually got refactored into a clearer form that better expressed the real problems at hand.

Now that we have a rich test suite to help protect correctness, I've shifted my attention to performance.  As before, eliminating unnecessary overhead has introduced a few awkward workarounds, but as I better understand what work is really needed (and conversely, what work can be avoided), I plan to continue looking for clearer ways to express the code.  And yes, I do expect some of that refactoring to yield ground on performance.

As long as we do continue to look for ways to improve the clarity of the code, I'm confident we'll find them.  And now that we've managed to hide the decoding and encoding internals within the library, I think we'll have the luxury post-1.0 of continuing to refactor our internal logic in fairly dramatic ways without impacting our clients.