@thomasvl 2017-03-21
Side comment -

One thing I've always sorta liked in the C++ impl is the `CodedInputStream`/`CodeOutputStream` models.  It allows the library to wrap something that writes to buffers, file descriptors, etc. meaning they can read/write directly with files, networking, etc. and not just in memory.

In our quest for performance, we've done all decoding/encoding around having the full payload in memory.  Which while ok for clients, might hurt things from a server usage pov.  I've got some ideas about how to speed out binary serialization, but they would never work in a streaming model, so I'm hesitant to actually put effort in to them.

Anyway, my fear is are we optimizing to the point that if we ever need to support streaming for server usage, we'll have to duplicate lots of code and/or accept what appears to be large performance hits because things were pushed too specific initially.

I haven't looked in detail, but the C++ has a few methods with _NoVirtual_ in the names and have very carefully crafted things so buffer only stuff can still be really fast, but still support the reading/writing external things.  There might be similar things we can do to split out specializations for in memory vs. more generic.